+++ app/src/sharedTest/java/org/oppia/app/recyclerview/BindableAdapterTest.kt
@@ -1,6 +1,5 @@
 package org.oppia.app.recyclerview
 
-import android.content.res.Resources
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
@@ -31,7 +30,6 @@ import androidx.test.espresso.matcher.BoundedMatcher
 import android.view.View
 import org.hamcrest.Description
 import org.hamcrest.Matcher
-import org.hamcrest.TypeSafeMatcher
 import org.oppia.app.databinding.TestTextViewForIntWithDataBindingBinding
 import org.oppia.app.databinding.TestTextViewForStringWithDataBindingBinding
 
@@ -327,56 +325,6 @@ class BindableAdapterTest {
       override fun matchesSafely(view: RecyclerView): Boolean {
         val viewHolder = view.findViewHolderForAdapterPosition(position) ?: return false
         return itemMatcher.matches(viewHolder.itemView)
-      }
-    }
-  }
-
-  /**
-   * This function returns a Matcher for an item inside RecyclerView from a specified position.
-   */
-  fun atPosition(recyclerViewId: Int, position: Int): Matcher<View> {
-    return atPositionOnView(recyclerViewId, position, -1)
-  }
-
-  /**
-   * This function returns a Matcher for a specific view within the item inside RecyclerView from a specified position.
-   */
-  fun atPositionOnView(recyclerViewId: Int, position: Int, targetViewId: Int): Matcher<View> {
-    return object : TypeSafeMatcher<View>() {
-      var resources: Resources? = null
-      var childView: View? = null
-
-      override fun describeTo(description: Description) {
-        var idDescription = Integer.toString(recyclerViewId)
-        if (this.resources != null) {
-          idDescription = try {
-            this.resources!!.getResourceName(recyclerViewId)
-          } catch (var4: Resources.NotFoundException) {
-            String.format(
-              "%s (resource name not found)",
-              recyclerViewId
-            )
-          }
-        }
-        description.appendText("with id: $idDescription")
-      }
-
-      public override fun matchesSafely(view: View): Boolean {
-        this.resources = view.resources
-        if (childView == null) {
-          val recyclerView = view.rootView.findViewById<View>(recyclerViewId) as RecyclerView
-          if (recyclerView.id == recyclerViewId) {
-            childView = recyclerView.findViewHolderForAdapterPosition(position)!!.itemView
-          } else {
-            return false
-          }
-        }
-        return if (targetViewId == -1) {
-          view === childView
-        } else {
-          val targetView = childView!!.findViewById<View>(targetViewId)
-          view === targetView
-        }
       }
     }
   }
