+++ app/src/sharedTest/java/org/oppia/app/topic/train/TopicTrainFragmentTest.kt
@@ -34,7 +34,7 @@ import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import androidx.test.rule.ActivityTestRule
-import org.oppia.app.recyclerview.BindableAdapterTest
+import org.oppia.app.recyclerview.RecyclerViewMatcher
 import org.oppia.app.topic.questionplayer.QuestionPlayerActivity
 
 /** Tests for [TopicTrainFragment]. */
@@ -42,7 +42,6 @@ import org.oppia.app.topic.questionplayer.QuestionPlayerActivity
 class TopicTrainFragmentTest {
 
   private var skillIdList = ArrayList<String>()
-  private val bindableAdapterTest:BindableAdapterTest = BindableAdapterTest()
 
   @get:Rule
   var activityTestRule: ActivityTestRule<TopicActivity> = ActivityTestRule(
@@ -59,7 +58,7 @@ class TopicTrainFragmentTest {
   fun testTopicTrainFragment_loadFragment_displaySkills_startButtonIsInactive() {
     ActivityScenario.launch(TopicActivity::class.java).use {
       onView(withId(R.id.master_skills_text_view)).check(matches(withText(R.string.topic_train_master_these_skills)))
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).check(matches(hasDescendant(withId(R.id.skill_check_box))))
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).check(matches(hasDescendant(withId(R.id.skill_check_box))))
       onView(withId(R.id.topic_train_start_button)).check(matches(not(isClickable())))
     }
   }
@@ -67,15 +66,15 @@ class TopicTrainFragmentTest {
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_isSuccessful() {
     ActivityScenario.launch(TopicActivity::class.java).use {
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,1)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,1)).perform(click())
     }
   }
 
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_startButtonIsActive() {
     ActivityScenario.launch(TopicActivity::class.java).use {
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
       onView(withId(R.id.topic_train_start_button)).check(matches(isClickable()))
     }
   }
@@ -83,16 +82,16 @@ class TopicTrainFragmentTest {
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_deselectSkills_isSuccessful() {
     ActivityScenario.launch(TopicActivity::class.java).use {
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
     }
   }
 
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_deselectSkills_startButtonIsInactive() {
     ActivityScenario.launch(TopicActivity::class.java).use {
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
-      onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
+      onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
       onView(withId(R.id.topic_train_start_button)).check(matches(not(isClickable())))
     }
   }
@@ -100,7 +99,7 @@ class TopicTrainFragmentTest {
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_clickStartButton_skillListTransferSuccessfully() {
     activityTestRule.launchActivity(null)
-    onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+    onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
     onView(withId(R.id.topic_train_start_button)).perform(click())
     intended(hasComponent(QuestionPlayerActivity::class.java.name))
     intended(hasExtra(QuestionPlayerActivity.getIntentKey(), skillIdList))
@@ -109,9 +108,9 @@ class TopicTrainFragmentTest {
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_configurationChange_skillsAreSelected() {
     activityTestRule.launchActivity(null)
-    onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+    onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
     activityTestRule.activity.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
-    onView(bindableAdapterTest.atPositionOnView(R.id.skill_recycler_view,0, R.id.skill_check_box)).check(matches(isChecked()))
+    onView(RecyclerViewMatcher.atPositionOnView(R.id.skill_recycler_view,0, R.id.skill_check_box)).check(matches(isChecked()))
   }
 
   @Test
@@ -125,7 +124,7 @@ class TopicTrainFragmentTest {
   @Test
   fun testTopicTrainFragment_loadFragment_selectSkills_configurationChange_startButtonRemainsActive() {
     activityTestRule.launchActivity(null)
-    onView(bindableAdapterTest.atPosition(R.id.skill_recycler_view,0)).perform(click())
+    onView(RecyclerViewMatcher.atPosition(R.id.skill_recycler_view,0)).perform(click())
     activityTestRule.activity.requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_LANDSCAPE
     onView(withId(R.id.topic_train_start_button)).check(matches(isClickable()))
   }
